+++ src/app/playlists/components/frame80_temp.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import { readableColor } from "polished";
 
-import DigCount from "./digcount";
+import DigCount from "./Digcount_temp";
 
 import { playArt } from "@/assets/images";
